Mask certain qstat/qacct errors when not all jobs are affected.
If qstat reports that a job is finished, but qacct output looks suspicious,
we try to give that job, and/or SGE, time to complete and/or recover.

We fall back to the previous behavior (tearing down the workflow) only if
every outstanding task has been affected by this SGE bug.