Defend against recycled SGE job ids.
If, during the course of an SGE upgrade, the qmaster service is re-set
to assign job id's from a new offset, which happens to overlap the tail
end of the previous range of job ids, then it's possible that qacct
will return metadata for two distinct jobs that share the same job id.
In pathological cases like that, we now return the metadata for the
most recently-completed job.