Call unexpected SGE output "spurious", not "corrupt".
During a recent outage, the wording in log messages confused a few
folks. The former _is_corrupt() method doesn't necessarily mean the
data are wrong, just that Cosmos either doesn't trust it or doesn't
know what to do with it. The method is now named more clearly,
_qacct_has_suspicious_stats(), and comments and log messages have
similarly been reworded.